+++ app/helpers/report_helper.rb
@@ -1,6 +1,6 @@
 module ReportHelper
   def report_download_link(report)
-    return report.export_url if report.export_url
+    return download_report_path(report) if report.export_filename
 
     report_path(report, format: :csv)
   end