+++ spec/helpers/report_helper_spec.rb
@@ -2,7 +2,7 @@ require "rails_helper"
 
 RSpec.describe ReportHelper, type: :helper do
   describe "#report_download_link" do
-    context "when the report doesn't have an export_url" do
+    context "when the report doesn't have an export_filename" do
       let(:report) { create(:report) }
 
       it "returns the report show path in CSV format" do
@@ -10,11 +10,11 @@ RSpec.describe ReportHelper, type: :helper do
       end
     end
 
-    context "when the report has an export_url" do
-      let(:report) { create(:report, export_url: "s3_link") }
+    context "when the report has an export_filename" do
+      let(:report) { create(:report, export_filename: "exported_csv") }
 
-      it "returns the export_url" do
-        expect(helper.report_download_link(report)).to eq("s3_link")
+      it "returns the report download path" do
+        expect(helper.report_download_link(report)).to eq(download_report_path(report))
       end
     end
   end