+++ CHANGELOG.md
@@ -1165,6 +1165,7 @@ activity and on its child transactions (which can be actuals, refunds, and adjus
 ## [unreleased]
 
 - Upload the report CSV to S3 when the report is approved
+- For approved reports, provide download of stored report CSV file instead of generating it from live data
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-126...HEAD
 [release-126]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-125...release-126